requirements.txt
requests==2.31.0
scikit-learn==1.4.0
coinbase-advanced-py==1.1.1
pandas~=2.2.0
joblib~=1.3.2
numpy~=1.26.4
tensorflow~=2.15.0
matplotlib~=3.8.3